Fix libsecret error on wsl. check if platform is wsl before require('keytar')

+++ grunt/removeToken.js
@@ -1,7 +1,6 @@
 const chalk = require('chalk');
 const path = require('path');
-const keytar = require('keytar')
-
+const isWsl = require('is-wsl');
 const {
 	get,
 } = require('lodash');
@@ -11,10 +10,9 @@ const removeToken = ( grunt ) => {
 
 	const modulePkg = grunt.file.readJSON( path.resolve( 'node_modules/wp-dev-env-grunt/package.json' ) );
 
-	const repoHost = getRepoHost( grunt );
-
-	if ( ! repoHost ) {
-		return new Promise( ( resolve, reject ) => () => {
+	return new Promise( ( resolve, reject ) => {
+		const repoHost = getRepoHost( grunt );
+		if ( ! repoHost ) {
 			grunt.log.writeln( 'Repository not hosted on ' + [
 				'GitHub',
 				'Bitbucket',
@@ -22,27 +20,26 @@ const removeToken = ( grunt ) => {
 				'Gitea',
 			].join( '||' ) );
 			reject();
-		} );
-	}
-
-	grunt.log.writeln( '' );
-
-	return new Promise( ( resolve, reject ) => {
-
-		keytar.deletePassword( modulePkg.name, get( repoHost, ['name'] ) ).then( res => {
-			grunt.log.writeln( res
-				? 'Successfully removed Personal Access Token for ' + chalk.yellow( get( repoHost, ['name'] ) )
-				: 'Couldn\'t remove Personal Access Token for ' + chalk.yellow( get( repoHost, ['name'] ) ) + ', entry not found'
-			);
-			resolve( res );
-		} ).catch( e => {
-			grunt.log.writeln( 'removeToken ....  what happend? Did you kill the process?' );
-			grunt.log.writeln( e );
-			reject( e );
-		} );
-
+		}
+		if ( isWsl ) {
+			grunt.log.writeln( 'Can not access credentials on wsl. Tokens can\'t be added, neiter removed.' );
+			reject();
+		} else {
+			grunt.log.writeln( '' );
+			const keytar = require('keytar');
+			keytar.deletePassword( modulePkg.name, get( repoHost, ['name'] ) ).then( res => {
+				grunt.log.writeln( res
+					? 'Successfully removed Personal Access Token for ' + chalk.yellow( get( repoHost, ['name'] ) )
+					: 'Couldn\'t remove Personal Access Token for ' + chalk.yellow( get( repoHost, ['name'] ) ) + ', entry not found'
+				);
+				resolve( res );
+			} ).catch( e => {
+				grunt.log.writeln( 'removeToken ....  what happend? Did you kill the process?' );
+				grunt.log.writeln( e );
+				reject( e );
+			} );
+		}
 	} );
-
 };
 
 module.exports = removeToken;
